requirements.txt
aioredis==2.0.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==5.0.0
asyncio-redis==0.16.0
asyncpg==0.30.0
bcrypt==4.1.2
build==1.2.1
CacheControl==0.14.0
cachetools==4.2.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.8
dnspython==2.7.0
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-filter==2.0.0
fastapi-login==1.10.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.19.1
filelock==3.13.3
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jose==1.0.0
keyring==24.3.1
makefun==1.15.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pwdlib==0.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-telegram-bot==13.7
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
redis==5.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
starlette==0.41.2
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.1
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==13.1
xattr==1.1.0
zipp==3.20.2
